% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KEADrugResponse_v0.12.R
\name{DrugResponse.pkgTest}
\alias{DrugResponse.pkgTest}
\title{test package}
\usage{
DrugResponse.pkgTest(x)
}
\arguments{
\item{x}{}
}
\description{
test if a package is installed.
}
\details{
This function reads cel file.
}
