% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KEADrugResponse_v0.12.R
\name{DrugResponse.readcel}
\alias{DrugResponse.readcel}
\title{read cel file}
\usage{
DrugResponse.readcel(path.data, patient_name)
}
\arguments{
\item{path.data}{}

\item{patient_name}{}
}
\value{
a vector of gene expression
}
\description{
Returns a vector of gene expression value for each patient.
}
\details{
This function reads cel file.
}
