% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KEADrugResponse_v0.12.R
\name{DrugResponse.score}
\alias{DrugResponse.score}
\title{calculate score}
\usage{
DrugResponse.score(test_index, drugwb_prob, drug_prob, test_geneExp,
  test_normalized, t_d, drug, test_geneExp_all, test_normalized_all)
}
\arguments{
\item{test_index}{patient index.}

\item{drugwb_prob}{}

\item{drug_prob}{}

\item{test_geneExp}{}

\item{test_normalized}{}

\item{t_d}{}

\item{drug}{}

\item{test_geneExp_all}{}

\item{test_normalized_all}{}

\item{test_normalized}{}
}
\value{
a positive or negtive score for each drug
}
\description{
Returns a positive or negtive response value for each drug.
}
\details{
This function calculate drug response score by using svm model vector.
}
